pyproject.toml
[tool.uv]
link-mode = "copy"  # Avoid hardlink warnings on Windows/OneDrive

[project]
name = "batch-summary-doubleword"
version = "0.1.0"
description = "Batch PDF summarization using Doubleword API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pypdf>=6.6.0",
    "pdfplumber>=0.11.9",
    "openai>=2.14.0",
    "python-dotenv>=1.1.0",
]
